修复播放wav时无法播放整数秒以后的数据问题以及加入适当帧延时使声音播放更流畅

+++ include/soundplayer.h
@@ -61,7 +61,7 @@ public:
 
 
     SoundStreamPos currentPos() const;
-    u64 currentSecond() const;
+    double currentSecond() const;
     void setPlayPos(const SoundStreamPos& pos);
     void setPlayPos(unsigned h, unsigned m, unsigned s) { setPlayPos(SoundStreamPos(h, m, s));}
